Move all sort code from table to column
