+++ packages/nimble-components/src/table-column/tests/types.ts
@@ -1,0 +1,8 @@
+export const ExampleSortType = {
+    firstColumnAscending: 'FirstColumnAscending',
+    firstColumnDescending: 'FirstColumnDescending',
+    firstColumnAscendingSecondColumnDescending:
+        'FirstColumnAscendingSecondColumnDescending'
+} as const;
+export type ExampleSortType =
+    (typeof ExampleSortType)[keyof typeof ExampleSortType];
