+++ packages/nimble-components/src/table/tests/table.stories.ts
@@ -2,9 +2,8 @@ import { html, ref } from '@microsoft/fast-element';
 import type { Meta, StoryObj } from '@storybook/html';
 import { withXD } from 'storybook-addon-xd-designs';
 import { createUserSelectedThemeStory, usageWarning } from '../../utilities/tests/storybook';
-import { ExampleDataType, ExampleSortType } from './types';
+import { ExampleDataType } from './types';
 import { Table, tableTag } from '..';
-import { TableColumnSortDirection } from '../types';
 import { iconUserTag } from '../../icons/user';
 import { menuTag } from '../../menu';
 import { menuItemTag } from '../../menu-item';
@@ -12,16 +11,11 @@ import { tableColumnTextTag } from '../../table-column/text';
 
 interface TableArgs {
     data: ExampleDataType;
-    sortedColumns: ExampleSortType;
     idFieldName: undefined;
     validity: undefined;
     checkValidity: undefined;
     tableRef: Table;
     updateData: (args: TableArgs) => void;
-    getColumnSortData: (
-        columnId: string,
-        args: TableArgs
-    ) => { direction: TableColumnSortDirection, index: number | undefined };
 }
 
 const simpleData = [
@@ -69,31 +63,6 @@ const dataSetIdFieldNames = {
     [ExampleDataType.largeDataSet]: 'id'
 } as const;
 
-const sortedOptions = {
-    [ExampleSortType.firstColumnAscending]: [
-        {
-            columnId: 'first-name-column',
-            sortDirection: TableColumnSortDirection.ascending
-        }
-    ],
-    [ExampleSortType.firstColumnDescending]: [
-        {
-            columnId: 'first-name-column',
-            sortDirection: TableColumnSortDirection.descending
-        }
-    ],
-    [ExampleSortType.firstColumnAscendingSecondColumnDescending]: [
-        {
-            columnId: 'first-name-column',
-            sortDirection: TableColumnSortDirection.ascending
-        },
-        {
-            columnId: 'last-name-column',
-            sortDirection: TableColumnSortDirection.descending
-        }
-    ]
-} as const;
-
 const overviewText = 'The `nimble-table` is a component that offers a way to render tabular data in a variety of ways in each column.';
 
 const dataDescription = `To set the data on the table, call \`setData()\` with an array data records. Each record is made up of fields,
@@ -112,10 +81,6 @@ The table will not automatically update if the contents of the array change afte
     Setting a new value on the property in Blazor will internally call \`setData()\` on the web component.
 </details>
 `;
-
-const sortedColumnsDescription = `A column within the table is configured to be sorted by specifying a \`sort-direction\` and a \`sort-index\` on
-it. The \`sort-direction\` indicates the direction to sort (\`ascending\` or \`descending\`), and the \`sort-index\` specifies the sort precedence
-of the column within the set of all sorted columns. Columns within the table will be sorted from lowest \`sort-index\` to highest \`sort-index\`.`;
 
 const idFieldNameDescription = `An optional string attribute that specifies the field name within a row's record to use as a row's ID.
 If the attribute is not specified, a default ID will be generated. If the attribute is invalid, no rows in the table will be rendered,
@@ -166,7 +131,6 @@ const metadata: Meta<TableArgs> = {
                 column-id="first-name-column"
                 field-name="firstName" placeholder="no value"
                 action-menu-slot="name-menu" action-menu-label="Configure name"
-                sort-direction="${x => x.getColumnSortData('first-name-column', x).direction}" sort-index="${x => x.getColumnSortData('first-name-column', x).index}"
             >
                 <${iconUserTag} title="First Name"></${iconUserTag}>
             </${tableColumnTextTag}>
@@ -174,14 +138,12 @@ const metadata: Meta<TableArgs> = {
                 column-id="last-name-column"
                 field-name="lastName" placeholder="no value"
                 action-menu-slot="name-menu" action-menu-label="Configure name"
-                sort-direction="${x => x.getColumnSortData('last-name-column', x).direction}" sort-index="${x => x.getColumnSortData('last-name-column', x).index}"
             >
                 Last Name
             </${tableColumnTextTag}>
             <${tableColumnTextTag}
                 column-id="favorite-color-column"
                 field-name="favoriteColor" placeholder="no value"
-                sort-direction="${x => x.getColumnSortData('favorite-color-column', x).direction}" sort-index="${x => x.getColumnSortData('favorite-color-column', x).index}"
             >
                 Favorite Color
             </${tableColumnTextTag}>
@@ -189,7 +151,6 @@ const metadata: Meta<TableArgs> = {
                 column-id="quote-column"
                 field-name="quote" placeholder="no value"
                 action-menu-slot="quote-menu" action-menu-label="Configure quote"
-                sort-direction="${x => x.getColumnSortData('quote-column', x).direction}" sort-index="${x => x.getColumnSortData('quote-column', x).index}"
             >
                 Quote
             </${tableColumnTextTag}>
@@ -221,22 +182,6 @@ const metadata: Meta<TableArgs> = {
                 }
             }
         },
-        sortedColumns: {
-            name: 'sort configuration',
-            description: sortedColumnsDescription,
-            options: Object.values(ExampleSortType),
-            control: {
-                type: 'radio',
-                labels: {
-                    [ExampleSortType.firstColumnAscending]:
-                        'First name ascending',
-                    [ExampleSortType.firstColumnDescending]:
-                        'First name descending',
-                    [ExampleSortType.firstColumnAscendingSecondColumnDescending]:
-                        'First name ascending then last name descending'
-                }
-            }
-        },
         idFieldName: {
             name: 'id-field-name',
             table: {
@@ -265,15 +210,9 @@ const metadata: Meta<TableArgs> = {
                 disable: true
             }
         },
-        getColumnSortData: {
-            table: {
-                disable: true
-            }
-        }
     },
     args: {
         data: ExampleDataType.simpleData,
-        sortedColumns: ExampleSortType.firstColumnAscending,
         idFieldName: undefined,
         validity: undefined,
         checkValidity: undefined,
@@ -286,29 +225,6 @@ const metadata: Meta<TableArgs> = {
                 x.tableRef.setData(dataSets[x.data]);
             })();
         },
-        getColumnSortData: (
-            columnId: string,
-            args: TableArgs
-        ): {
-            direction: TableColumnSortDirection,
-            index: number | undefined
-        } => {
-            const sortData = sortedOptions[args.sortedColumns];
-            const matchingIndex = sortData.findIndex(
-                sortedColumn => sortedColumn.columnId === columnId
-            );
-            if (matchingIndex === -1) {
-                return {
-                    direction: TableColumnSortDirection.none,
-                    index: undefined
-                };
-            }
-
-            return {
-                direction: sortData[matchingIndex]!.sortDirection,
-                index: matchingIndex
-            };
-        }
     }
 };
