+++ packages/nimble-components/src/table/tests/types.ts
@@ -4,12 +4,3 @@ export const ExampleDataType = {
 } as const;
 export type ExampleDataType =
     (typeof ExampleDataType)[keyof typeof ExampleDataType];
-
-export const ExampleSortType = {
-    firstColumnAscending: 'FirstColumnAscending',
-    firstColumnDescending: 'FirstColumnDescending',
-    firstColumnAscendingSecondColumnDescending:
-        'FirstColumnAscendingSecondColumnDescending'
-} as const;
-export type ExampleSortType =
-    (typeof ExampleSortType)[keyof typeof ExampleSortType];
